Port assessment reordering js to webpack and use es6
